+++ src/assets/js/components/ExerciseList.vue
@@ -1,12 +1,14 @@
 <template>
   <div class="container">
     <h1>Exercises</h1>
+    <button @click="addNewExercise()">+</button>
     <div
       v-for="(exercises, equipment) in exercisesByEquipment"
       :key="equipment"
       class="flex-container"
     >
-      <h3>{{ equipment }}</h3>
+      <h3 v-if="equipment">{{ equipment }}</h3>
+      <h3 v-else>Uncategorized</h3>
       <ExerciseListItem
         v-for="exercise in exercises"
         :key="exercise.id"
@@ -20,7 +22,7 @@
 import ExerciseListItem from "./ExerciseListItem";
 import { createNamespacedHelpers } from "vuex";
 
-const { mapState } = createNamespacedHelpers("exercises");
+const { mapState, mapActions } = createNamespacedHelpers("exercises");
 
 export default {
   name: "ExerciseList",
@@ -38,6 +40,15 @@ export default {
         result[exercise.equipment].push(exercise);
       });
       return result;
+    }
+  },
+  methods: {
+    ...mapActions(["addExercise"]),
+    addNewExercise: function() {
+      var router = this.$router;
+      this.addExercise().then(function(id) {
+        router.push({ path: "/exercises/" + id });
+      });
     }
   }
 };