+++ src/assets/js/store/modules/exercises/actions.js
@@ -2,7 +2,13 @@ export default {
   handler() {},
 
   addExercise(context, payload) {
-    context.commit("addExercise", payload);
+    let newExercise = payload || {
+      id: this.generateId(),
+      name: "New Exercise",
+      equipment: ""
+    };
+    context.commit("addExercise", newExercise);
+    return newExercise.id;
   },
 
   removeExercise(context, payload) {