+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Workouts are now editable
+- Exercises are now editable
 
 ## 0.1.0 - 2018-12-06